requirements.txt
alembic==1.7.7
Brotli==1.0.9
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
email-validator==1.2.1
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Compress==1.12
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
plotly==5.7.0
psycopg==3.0.15
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
tenacity==8.0.1
tzdata==2022.1
Werkzeug==2.1.2
WTForms==3.0.1
